requirements.txt
discord.py
python-telegram-bot
requests